refactor trial balance report components for improved structure and maintainability
- extract print styles into a separate PrintStyles file
- create a dedicated TrialBalanceTable component for rendering the trial balance table
- implement a custom hook, useTrialBalance, to manage trial balance state and fetching logic
